Make reservation state persist on navigation
