+++ src/pages/Rockets.js
@@ -1,21 +1,10 @@
-import React, { useEffect } from 'react';
-import { useDispatch, useSelector } from 'react-redux';
-import fetchRockets from '../api/rocketApiCall';
+import React from 'react';
+import { useSelector } from 'react-redux';
 import RocketContainer from '../components/RocketContainer';
-import { getRockets } from '../redux/rockets/rockets';
 
 function Rockets() {
-  const dispatch = useDispatch();
-
-  useEffect(() => {
-    async function data() {
-      const rockets = await fetchRockets();
-      dispatch(getRockets(rockets));
-    }
-    data();
-  }, []);
-
   const rockets = useSelector((state) => state.rockets);
+
   return (
     <>
       {rockets.map((rocket) => (
@@ -25,6 +14,7 @@ function Rockets() {
           name={rocket.rocket_name}
           description={rocket.description}
           image={rocket.flickr_image}
+          reserved={rocket.reserved}
         />
       ))}
     </>
